Fixes EGD's gain scheduling config when absent
Summary:
Fixes issue with EGD's gain scheduling mode configuration when the drive's pre-loaded mode is desired. The root of the problem was a type cast issue between an enumeration type and an integer type.

Test Plan:
Tested with fcat and an actual motor. The fcat ROS1 module was modified for this test to support the EGD Fastcat driver.

Reviewers: JosephBowkett, alex-brinkman